fix(plugin-sdk): keep session isolation CI-clean
